3989 - Topic/Footer: hide resolving and sending messages when in review
- Introduce new component Topic/Footer
- Move most of the logic in the footer and clean up
